#347 fixed

+++ juntagrico/templatetags/depot_extras.py
@@ -32,11 +32,11 @@ def count_units(subs_or_types):
     elif isinstance(subs_or_types, QuerySet):
         if subs_or_types.model is Subscription:
             # case 2: sum each unit of each subscription type
-            units = {'units': str(sum([int(sub.active_parts.aggregate(units=Sum('type__size__units'))['units'] or 0) for sub in subs_or_types.all()]))}
+            units = {'units': str(sum([float(sub.active_parts.aggregate(units=Sum('type__size__units'))['units'] or 0) for sub in subs_or_types.all()]))}
         elif subs_or_types.model is SubscriptionType:
             # case 3: queryset of types is passed
             units = subs_or_types.aggregate(units=Sum('size__units'))
-    return int(units['units'] or 0)
+    return float(units['units'] or 0)
 
 
 @register.filter
